backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
clarifai==10.0.0
Pillow==10.1.0
boto3==1.29.7
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
